Renamed lab2.py and test_lab2.py files
