+++ src/min_eating_time.py
